frame-sorter: Fix wrong type conversion
It was putting a uint64_t timestamp into a double, losing precision and
not de-duplicating correctly.

+++ lib/frame_sorter_impl.h
@@ -16,7 +16,8 @@ namespace iridium {
 class frame
 {
 public:
-    double timestamp;
+    // Needs to be a signed integer for std::abs to be available
+    int64_t timestamp;
     double center_frequency;
 
     friend bool operator<(const frame& l, const frame& r)
